Consolidate diff-cache shape check via DiffCacheKey::isCurrentShape
Three call sites duplicated the marker-key invariant: LoadFileDiffAction
ran the full key chain, BuildDiffContextAction and diff-file hydrate
each replicated a short `lineTypesAreEnum` probe. Extract a single
predicate so all three readers share the same source of truth and the
next format-change marker is added in one place.
