+++ app/Support/DiffCacheKey.php
@@ -15,4 +15,22 @@ final class DiffCacheKey
     {
         return 'rfa_diff_v8_'.hash('xxh128', $projectIdOrRepoPath.':'.$contextKey.':'.$fileId);
     }
+
+    /**
+     * Each historical shape change adds a marker key; missing any means the
+     * entry predates a format change and must be recomputed.
+     *
+     * @phpstan-assert-if-true array<string, mixed> $cached
+     */
+    public static function isCurrentShape(mixed $cached): bool
+    {
+        return is_array($cached)
+            && array_key_exists('syntaxStyles', $cached)
+            && array_key_exists('isSymlink', $cached)
+            && array_key_exists('tableAligned', $cached)
+            && array_key_exists('newFileLineCount', $cached)
+            && array_key_exists('headingsAnnotated', $cached)
+            && array_key_exists('gridLayout', $cached)
+            && array_key_exists('lineTypesAreEnum', $cached);
+    }
 }
